Updated updateConfigBundle script to latest version 0.5.5 with hack that ignores local version notes file

+++ scripts/updateConfigBundle.groovy
@@ -24,6 +24,7 @@
  *          0.5.2: Fix local plugin directory creation
  *          0.5.3: Delete folder accidentally created with version 0.5.1 when trying to write jar file
  *          0.5.4: Update autotext merging logic
+ *          0.5.5: Added hack to run the full customization each time (only to be in the version that does not autoLaunch)
  *
  */
 
@@ -169,7 +170,7 @@ localPrefFile = new File(confDir + File.separator + "omegat.prefs")
 bundleAutoText = new File(tmpConfigDir.toString() + File.separator + "omegat.autotext")
 localAutoText = new File(confDir + File.separator + "omegat.autotext")
 // A hack to run the full customization each time
-if (verFile.exists()) {
+if (verFile.exists() && !autoLaunch) {
     verFile.delete()
 }
